+++ src/components/LicenseAmendmentList/LicenseAmendmentList.js
@@ -1,0 +1,66 @@
+import React from 'react';
+import PropTypes from 'prop-types';
+import { FormattedMessage } from 'react-intl';
+import Link from 'react-router-dom/Link';
+
+import { InfoPopover, MultiColumnList } from '@folio/stripes/components';
+import { LicenseEndDate } from '@folio/stripes-erm-components';
+
+import FormattedUTCDate from '../FormattedUTCDate';
+import { urls } from '../utilities';
+import css from './LicenseAmendmentList.css';
+
+export default class LicenseAmendmentList extends React.Component {
+  static propTypes = {
+    amendments: PropTypes.arrayOf(PropTypes.shape({
+      id: PropTypes.string.isRequired,
+      name: PropTypes.string.isRequired,
+      note: PropTypes.string,
+      startDate: PropTypes.string,
+      status: PropTypes.shape({
+        label: PropTypes.string,
+      }),
+    })),
+    id: PropTypes.string,
+    license: PropTypes.shape({
+      id: PropTypes.string.isRequired,
+    }),
+    renderStatuses: PropTypes.bool,
+  }
+
+  render() {
+    const {
+      amendments,
+      id,
+      license,
+      renderStatuses,
+    } = this.props;
+
+    return (
+      <MultiColumnList
+        columnMapping={{
+          note: '',
+          name: <FormattedMessage id="ui-agreements.license.amendment" />,
+          status: <FormattedMessage id="ui-agreements.status" />,
+          startDate: <FormattedMessage id="ui-agreements.license.prop.startDate" />,
+          endDate: <FormattedMessage id="ui-agreements.license.prop.endDate" />,
+        }}
+        contentData={amendments}
+        formatter={{
+          note: a => (a.note ? <InfoPopover contentClass={css.note} content={a.note} /> : ''),
+          name: a => <Link to={urls.amendmentView(license.id, a.id)}>{a.name}</Link>,
+          status: a => (a.status ? a.status.label : '-'),
+          startDate: a => (a.startDate ? <FormattedUTCDate value={a.startDate} /> : '-'),
+          endDate: a => <LicenseEndDate license={a} />,
+        }}
+        id={id}
+        interactive={false}
+        visibleColumns={
+          renderStatuses ?
+            ['note', 'name', 'status', 'startDate', 'endDate'] :
+            ['note', 'name', 'startDate', 'endDate']
+        }
+      />
+    );
+  }
+}