+++ src/constants/index.js
@@ -1,3 +1,4 @@
 export { default as entityTypes } from './entityTypes';
 export { default as entityTypeTranslationKeys } from './entityTypeTranslationKeys';
 export { default as entityTypePluralizedTranslationKeys } from './entityTypePluralizedTranslationKeys';
+export { default as statuses } from './statuses';