+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 * Move the Save & close button and add a Cancel button to Pane Footer. ERM-411.
 * Apply the new large headline design to Agreements. ERM-260.
 * Display coverage for Agreement lines on Create Agreement pane. ERM-168.
+* Added ability to manage amendments for linked licenses. ERM-428 ERM-429
 
 ## 3.2.4 2019-09-10
 * Add support for interface `invoice` version `2.0`.